+++ src/app/search/search-ingredient-folder.tsx
@@ -3,16 +3,17 @@ import { IngredientCategoryType } from '@/src/types/ingredient-category';
 
 interface SearchIngredientFolderProps {
     category: IngredientCategoryType;
+    folded: boolean;
     onChange: () => void;
 }
 
-export function SearchIngredientFolder({ category, onChange }: SearchIngredientFolderProps) {
+export function SearchIngredientFolder({ category, folded, onChange }: SearchIngredientFolderProps) {
     const ingredientId = `${category}-folding`;
 
     return (
         <span className={styles['search-ingredient-category__ingredient']} key={ingredientId}>
             <input type="checkbox" name={ingredientId} id={ingredientId} className="d-none" onChange={onChange} />
-            <label htmlFor={ingredientId}>Fold/Unfold</label>
+            <label htmlFor={ingredientId}>{folded ? 'Rozwiń' : 'Zwiń'}</label>
         </span>
     );
 }